add google analytics rename migrations fixe named_scope to page add fields to scaffolds add helper application changed css to messages add js application add component upload and image

+++ views/admin/scaffolds/index.ctp
@@ -4,10 +4,8 @@
 	<table>
 		<tr>
 			<?php foreach ($schema as $field => $options):?>
-				<?php if (in_array($field, $except)): ?>
-					<?php continue; ?>
-				<?php endif; ?>
-				<th><?php echo $field;?></th>
+				<?php if (in_array($field, $except)){continue;} ?>
+				<th><?php echo __(Inflector::humanize(preg_replace('/_id$/', '', $field)));?></th>
 			<?php endforeach;?>
 		</tr>
 		<?php foreach ($collection as $obj): ?>
@@ -15,20 +13,34 @@
 				<?php
 					foreach ($schema as $field => $options){
 						if(in_array($field, $except)){ continue; }
-					
-						switch ($options["type"]) {
-							case 'date':
-								echo "\t\t\t\t<td>{$time->format(Configure::read('format.date'), $obj[$model][$field])}</td>\n";
-								break;
-							case 'datetime':
-								echo "\t\t\t\t<td>{$time->format(Configure::read('format.datetime'), $obj[$model][$field])}</td>\n";
-								break;
-							case 'text':
-								echo "\t\t\t\t<td>{$text->truncate(strip_tags($obj[$model][$field]))}</td>\n";
-								break;
-							default:
-								echo "\t\t\t\t<td>{$obj[$model][$field]}</td>\n";
-								break;
+						
+						$isKey = false;
+						
+						if (!empty($associations['belongsTo'])) {
+							foreach ($associations['belongsTo'] as $_alias => $_details) {
+								if ($field === $_details['foreignKey']) {
+									$isKey = true;
+									echo "\t\t\t\t<td>" . $obj[$_alias][$_details['displayField']] . "</td>\n";
+									break;
+								}
+							}
+						}
+						if ($isKey != true)
+						{
+							switch ($options["type"]) {
+								case 'date':
+									echo "\t\t\t\t<td>{$time->format(Configure::read('format.date'), $obj[$model][$field])}</td>\n";
+									break;
+								case 'datetime':
+									echo "\t\t\t\t<td>{$time->format(Configure::read('format.datetime'), $obj[$model][$field])}</td>\n";
+									break;
+								case 'text':
+									echo "\t\t\t\t<td>{$text->truncate(strip_tags($obj[$model][$field]))}</td>\n";
+									break;
+								default:
+									echo "\t\t\t\t<td>{$obj[$model][$field]}</td>\n";
+									break;
+							}
 						}
 					}
 				?>
